+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageContainerView.swift
@@ -60,7 +60,7 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
 
     public var body: some View {
         HStack(alignment: .bottom) {
-            if message.type == .system || message.type == .error {
+            if message.type == .system || (message.type == .error && message.isBounced == false) {
                 factory.makeSystemMessageView(message: message)
             } else {
                 if message.isRightAligned {
@@ -105,7 +105,7 @@ public struct MessageContainerView<Factory: ViewFactory>: View {
                                 )
                                 : nil
 
-                            message.localState == .sendingFailed ? SendFailureIndicator() : nil
+                            (message.localState == .sendingFailed || message.isBounced) ? SendFailureIndicator() : nil
                         }
                     )
                     .background(